% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_validation}
\alias{check_validation}
\title{Check validation results}
\usage{
check_validation(out)
}
\arguments{
\item{out}{output from \code{\link[validate:confront]{validate::confront()}}.}
}
\value{
\code{FALSE} if the validation had any failures or resulted in any errors or warnings,
otherwise \code{TRUE}, invisibly.
}
\description{
Check results from \code{\link[validate:confront]{validate::confront()}} and raise warnings if any validations had failures, or
if any validations resulted in an error or a warning.
}
