% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{gmean}
\alias{gmean}
\title{Geometric Mean}
\usage{
gmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric or complex vector.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}
}
\value{
Geometric mean, \verb{exp(mean(log(x))}. \code{NA} values are removed if \code{na.rm} is \code{TRUE}.
}
\description{
Function for the geometric mean.
}
