% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{quantrr-package}
\alias{quantrr}
\alias{quantrr-package}
\title{quantrr: Quantify and Report (Technology) Risk}
\description{
A streamlined tool designed to make it easier to quantify risk using expert estimation. Given expert estimates for loss frequency and magnitude, generate a report that forecasts losses by risk and in aggregate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/quantrr/}
  \item \url{https://github.com/jabenninghoff/quantrr}
  \item Report bugs at \url{https://github.com/jabenninghoff/quantrr/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
