% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widget_estimates}
\alias{widget_estimates}
\title{Widget System Estimates}
\format{
A data frame with 21 rows and 6 columns:
\describe{
\item{Risk}{Risk short name from "Risks" tab}
\item{Expert}{Name of Subject Matter Expert (SME) who provided the estimate}
\item{Frequency (per yer)}{Loss Event Frequency estimate}
\item{Low (5\%)}{5th percentile estimate of Loss Event Magnitude}
\item{High (95\%)}{95th percentile estimate of Loss Event Magnitude}
\item{Most Likely}{50th percentile estimate of Loss Event Magnitude (median)}
}
}
\usage{
widget_estimates
}
\description{
Data from the "Estimates" tab of the Widget System example spreadsheet, imported with
\code{\link[readxl:read_excel]{readxl::read_xlsx()}}. Contains estimates on each of 3 risks from 7 different experts,
including blanks (\code{NA}) for estimates not provided by an expert.
}
\keyword{datasets}
