% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widget_risks}
\alias{widget_risks}
\title{Widget System Risks}
\format{
A data frame with 3 rows and 2 columns:
\describe{
\item{Risk}{Risk short name}
\item{Description}{Risk long description}
}
}
\usage{
widget_risks
}
\description{
Data from the "Risks" tab of the Widget System example spreadsheet, imported with
\code{\link[readxl:read_excel]{readxl::read_xlsx()}}. Contains 3 risks.
}
\keyword{datasets}
